Fixed a bug in the year calculation
I accidentally overwrote ix (the number of days since the epoch) by the year
number, which caused problems.

Also, fixed a bug in daysSinceEpoch (this function didn't preserve de and bc).

Finally, modified the code to actually return the computed year in ix. All
tests pass now for this function (for the year only - the month and day are not
yet computed).